Move plugins installation to init container
Install jenkins plugins in the init container as requested in
https://github.com/jenkinsci/kubernetes-operator/issues/337
Changes introduced in this commit make jenkins-master container
sustainable to readiness/liveness probe failures when downloading
plugins, as plugins are downloaded in separate container (initContainer,
jenkins-init).

Operations with JENKINS_REF and REF, introduced in this commit are
required in order to properly handle custom jenkins images with
pre-installed plugins. Simply using REF path as EmptyDir is not desired
in this case as it will mask all pre-installed plugins in the image and
thus make such `cached` image useless. In order to support such
scenario, a different volume is created, which is later populated with
contents of REF from the image and only afterwards plugins are
installed.

Closes https://github.com/jenkinsci/kubernetes-operator/issues/337