Parameterised liveness and readiness probe in helm chart
The jenkins master config accepted by the operator has the liveness and
readiness probe already parameterised. This however has been hard coded
in the helm chart implementation for this operator. I have just moved
the values to be set using variables from values.yaml file. I have also
added an enable option so that the user of the chart may implement the
probes as per their volition.

Additionally this comes in handy if the no of plugins to be installed is
high and causes the jenkins master to terminate immaturely due to probe
timeout. Ideally the plugins must be installed via an init container but
given that the pr #393 to implement that is in stale state, I have
raised this pr to be a workaround to this issue.