Remove the validation of plugin version against a regex
The plugin versions have been validated against a regex. This has
led to an issue where the version formats which are allowed by
the Jenkins plugin repositories have apparently changed, which led to
failures to install newer versions of some plugins.

It was necessary to remove the plugin version validation in order to:
- Prevent unnecessary failures in Jenkins Operator when versions which
  are allowed by Jenkins plugin repositories change
- Enable more transparent error diagnosing, since if the version
  of plugin is not allowed, then Jenkins instance will fail to start
  with a clear message, stating that it was not able to find the plugin